Created UDI test functions to build test data and fixed UDI debugging tools.
* The test functions build a suite of grid models testing each of the udi features.
* When creating a udi the header id is passed through to the octree structure.
* Updated GenPointCloud to take a new parameter "createTestUDIs".  When this is passed it outputs the test models.
* Fixed Win32 warnings in udEditorModel.cpp

+++ udPlatform/Source/udPlatform.cpp
@@ -407,14 +407,14 @@ void udMemoryDebugLogMemoryStats()
 {
   udDebugPrintf("Memory Stats\n");
 
-  uint32_t totalMemory = 0;
+  size_t totalMemory = 0;
   for (MemTrackMap::iterator memIt = pMemoryTrackingMap->begin(); memIt != pMemoryTrackingMap->end(); ++memIt)
   {
     const MemTrack &track = memIt->second;
     totalMemory += track.size;
   }
 
-  udDebugPrintf("Total allocated Memory %d\n", totalMemory);
+  udDebugPrintf("Total allocated Memory %llu\n", totalMemory);
 }
 
 #else
